requirements/base.txt
numpy>=1.13.3
pandas>=0.23.1
skorch>=0.2.0
matplotlib>=2.2.2
munch>=2.3.2
sacred>=0.7.3
pyyaml>=3.12
feather-format>=0.4.0
